requirements.txt
django>=5.0,<6.0
djangorestframework
django-cors-headers
pillow
python-decouple
psycopg2-binary
dj-database-url
python-dotenv
cloudinary
django-cloudinary-storage
